requirements.txt
# Core Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client (for inter-service communication)
httpx==0.26.0

# Caching & Sessions
redis==5.0.1

# Kafka (for event-driven architecture)
aiokafka==0.10.0

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
